more tweaks based on comm. feedback
